test: factorize runCli
