+++ libs/ledger-live-common/src/e2e/speculos.ts
@@ -53,6 +53,7 @@ import { sanitizeError } from ".";
 import { sendVechain } from "./families/vechain";
 import { getDeviceCoordinates } from "./deviceCoordinates";
 import { sendInternetComputer } from "./families/internet_computer";
+import { sleep } from "./index";
 
 const isSpeculosRemote = process.env.REMOTE_SPECULOS === "true";
 
@@ -566,7 +567,7 @@ export async function waitFor(text: string, maxAttempts = 60): Promise<string> {
       return texts;
     }
 
-    await waitForTimeOut(500);
+    await sleep(500);
   }
 
   throw new Error(
@@ -591,7 +592,7 @@ export async function waitForReviewTransaction(): Promise<void> {
       await waitFor(DeviceLabels.REVIEW_TRANSACTION);
       return;
     }
-    await waitForTimeOut(500);
+    await sleep(500);
   }
 }
 
@@ -665,7 +666,7 @@ export const pressUntilTextFound = withDeviceController(
         } else {
           await buttons.right();
         }
-        await waitForTimeOut(200);
+        await sleep(200);
       }
 
       const screensLog = [...seenScreens].map((s, i) => `[${i + 1}] "${s}"`).join(" → ");
@@ -702,10 +703,6 @@ export async function takeScreenshot(port?: number): Promise<Buffer | undefined>
   } catch (error) {
     console.error("Error downloading speculos screenshot:", sanitizeError(error));
   }
-}
-
-export async function waitForTimeOut(ms: number) {
-  return new Promise(resolve => setTimeout(resolve, ms));
 }
 
 export const removeMemberLedgerSync = withDeviceController(
